+++ CHANGES.md
@@ -7,7 +7,10 @@ Fast and flexible PHP library rendering plain ASCII tables.
 ## Changes
 
 * dev
-  * Main class name changed to `TextTable`. 
+  * Main class name changed to `TextTable`.
+  * `RendererContract`'s `render()` no longer expects returns `string[]`.
+  * Writers and `WriterContract` are completely gone as they are no longer needed.
+
 
 * 1.7.0 (2022-10-19)
   * Corrected rendering tables with first/last column hidden.